create redux foler and split reducer file into files for each reducer
