+++ client/src/redux/reducers/flagReducer.js
@@ -1,0 +1,10 @@
+function reducer(state = 0, action) {
+  switch (action.type) {
+    case "REFRESH_FLAG":
+      return !state;
+    default:
+      return state;
+  }
+}
+
+export default reducer;
