+++ client/src/redux/reducers/index.js
@@ -1,0 +1,7 @@
+import counterReducer from "./counterReducer";
+import scoreReducer from "./scoreReducer";
+import percentReducer from "./percentReducer";
+import flagReducer from "./flagReducer";
+import instructionsCounterReducer from "./instuctionsCounterReducer";
+
+export { counterReducer, scoreReducer, percentReducer, flagReducer, instructionsCounterReducer };
