+++ client/src/redux/reducers/isCompletedReducer.js
@@ -1,0 +1,21 @@
+// const initialState = localStorage.getItem("complete") || false;
+
+// function reducer(state = initialState, action) {
+//   switch (action.type) {
+//     case "IS_COMPLETED": {
+//       return action.isCompleted;
+//     }
+//     case "RESTART":
+//       return {
+//         inattentionScore: 0,
+//         hyperactivityScore: 0,
+//         totalScore: 0,
+//         isCompleted: false,
+//         score: []
+//       };
+//     default:
+//       return state;
+//   }
+// }
+
+// export default reducer;
