+++ client/src/redux/reducers/percentReducer.js
@@ -1,0 +1,10 @@
+function reducer(state = 0, action) {
+  switch (action.type) {
+    case "ADD_PERCENT":
+      return action.percent;
+    default:
+      return state;
+  }
+}
+
+export default reducer;
