+++ client/src/redux/reducers/totalScoreReducer.js
@@ -1,0 +1,21 @@
+// const initialState = {
+//   inattentionScore: localStorage.getItem("inattentionScore") || 0,
+//   hyperactivityScore: localStorage.getItem("hyperactivityScore") || 0,
+//   totalScore: localStorage.getItem("totalScore") || 0
+// };
+
+// function reducer(state = initialState, action) {
+//   switch (action.type) {
+//     case "TOTAL_SCORE":
+//       return {
+//         ...state,
+//         inattentionScore: action.inattentionScore,
+//         hyperactivityScore: action.hyperactivityScore,
+//         totalScore: action.totalScore
+//       };
+//     default:
+//       return state;
+//   }
+// }
+
+// export default reducer;
